+++ test/Armada.Test.Database/DatabaseTestRunner.cs
@@ -155,6 +155,8 @@ namespace Armada.Test.Database
             await RunTest("Skill_Create_Read_Update_Reopen_Window_Delete", "Operational", () => roundTrips.VerifySkillsAsync(token), token);
             await RunTest("ProjectProfile_Create_Read_Update_Reopen_Window_Delete", "Operational", () => roundTrips.VerifyProjectProfilesAsync(token), token);
             await RunTest("Playbook_Create_Read_And_Empty_Text_Reads_As_Null", "Operational", () => roundTrips.VerifyPlaybookEmptyTextAsync(token), token);
+            await RunTest("LandingJob_Create_Read_Update_Reopen_State_Delete", "Operational", () => roundTrips.VerifyLandingJobsAsync(token), token);
+            await RunTest("JudgeFollowUp_Upsert_Associate_Audit_Update_Reopen", "Operational", () => roundTrips.VerifyJudgeFollowUpsAsync(token), token);
             await RunTest("Pipeline_Update_And_Delete_Roll_Back_On_Failure", "Operational", () => TestPipelineWriteAtomicityAsync(token), token);
             await RunTest("RequestHistory_Timestamp_RoundTrip_And_Same_Day_Range", "Operational", () => TestRequestHistorySameDayRangeAsync(token), token);
             await RunTest("CheckRun_Create_Read_Update_Enumerate", "Operational", () => TestCheckRunCrudAsync(token), token);